Fix OOBE
