+++ src/main/java/me/noramibu/bettershulkers/recipe/BetterShulkersRecipes.java
@@ -7,7 +7,7 @@ import net.minecraft.registry.Registry;
 import net.minecraft.util.Identifier;
 
 public class BetterShulkersRecipes {
-    public static final RecipeSerializer<ShulkerUpgradeRecipeV2> SHULKER_UPGRADE_SERIALIZER = new ShulkerUpgradeRecipeV2.Serializer(ShulkerUpgradeRecipeV2::new);
+    public static final RecipeSerializer<ShulkerUpgradeRecipe> SHULKER_UPGRADE_SERIALIZER = new ShulkerUpgradeRecipe.Serializer(ShulkerUpgradeRecipe::new);
 
     public static void register() {
         Registry.register(Registries.RECIPE_SERIALIZER, Identifier.of(BetterShulkers.MOD_ID, "shulker_upgrade"), SHULKER_UPGRADE_SERIALIZER);
