+++ src/main/java/me/noramibu/bettershulkers/recipe/ShulkerUpgradeRecipe.java
@@ -1,79 +1,84 @@
 package me.noramibu.bettershulkers.recipe;
 
-import com.mojang.serialization.MapCodec;
-import me.noramibu.bettershulkers.BetterShulkers;
+import eu.pb4.polymer.core.api.item.PolymerRecipe;
+import eu.pb4.polymer.core.api.utils.PolymerObject;
 import me.noramibu.bettershulkers.util.ShulkerUtil;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.Items;
 import net.minecraft.network.RegistryByteBuf;
 import net.minecraft.network.codec.PacketCodec;
-import net.minecraft.recipe.CraftingRecipe;
-import net.minecraft.recipe.RecipeSerializer;
+import net.minecraft.recipe.*;
 import net.minecraft.recipe.book.CraftingRecipeCategory;
 import net.minecraft.recipe.input.CraftingRecipeInput;
 import net.minecraft.registry.RegistryWrapper;
+import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.world.World;
+import org.jetbrains.annotations.Nullable;
 
-public class ShulkerUpgradeRecipe implements CraftingRecipe {
+public class ShulkerUpgradeRecipe extends SpecialCraftingRecipe implements PolymerRecipe {
 
-    @Override
-    public boolean matches(CraftingRecipeInput inventory, World world) {
-        if (inventory.getWidth() != 3 || inventory.getHeight() != 3) {
-            return false;
-        }
-
-        if (!ShulkerUtil.isShulkerBox(inventory.getStackInSlot(1, 1))) {
-            return false;
-        }
-
-        if (!inventory.getStackInSlot(0, 0).isOf(Items.OBSIDIAN) ||
-            !inventory.getStackInSlot(0, 2).isOf(Items.OBSIDIAN) ||
-            !inventory.getStackInSlot(2, 0).isOf(Items.OBSIDIAN) ||
-            !inventory.getStackInSlot(2, 2).isOf(Items.OBSIDIAN)) {
-            return false;
-        }
-
-        Item material = null;
-        for (int row = 0; row < 3; row++) {
-            for (int col = 0; col < 3; col++) {
-                if ((row == 1 && col == 1) || (row % 2 == 0 && col % 2 == 0)) {
-                    continue;
-                }
-                ItemStack currentStack = inventory.getStackInSlot(row, col);
-                if (currentStack.isEmpty()) {
-                    return false;
-                }
-                if (material == null) {
-                    material = currentStack.getItem();
-                } else {
-                    if (!currentStack.isOf(material)) {
-                        return false;
-                    }
-                }
-            }
-        }
-        return material != null;
+    public ShulkerUpgradeRecipe(CraftingRecipeCategory category) {
+        super(category);
     }
 
     @Override
-    public ItemStack craft(CraftingRecipeInput inventory, RegistryWrapper.WrapperLookup lookup) {
-        ItemStack materialStack = inventory.getStackInSlot(0, 1);
-        ItemStack shulkerStack = inventory.getStackInSlot(1, 1).copy();
-        shulkerStack.setCount(1);
-        String materialId = ShulkerUtil.getItemId(materialStack);
-        ShulkerUtil.setShulkerMaterial(shulkerStack, materialId);
+    public boolean matches(CraftingRecipeInput input, World world) {
+        // Prevent OutOfBoundsException
+        // Also acts as a super quick escape path
+        if (input.getSize() != 9) {
+            return false;
+        }
+
+        // Check center for a shulker
+        if (!ShulkerUtil.isShulkerBox(input.getStackInSlot(1, 1))) {
+            return false;
+        }
+
+        // Check 4 corners for obsidian
+        if (!input.getStackInSlot(0, 0).isOf(Items.OBSIDIAN) ||
+                !input.getStackInSlot(0, 2).isOf(Items.OBSIDIAN) ||
+                !input.getStackInSlot(2, 0).isOf(Items.OBSIDIAN) ||
+                !input.getStackInSlot(2, 2).isOf(Items.OBSIDIAN)) {
+            return false;
+        }
+
+        // Check materials
+        Item materialItem = input.getStackInSlot(1, 0).getItem();
+        // Quick escape
+        if (materialItem == Items.AIR) {
+            return false;
+        }
+
+        return input.getStackInSlot(0, 1).isOf(materialItem)  &&
+                input.getStackInSlot(2, 1).isOf(materialItem) &&
+                input.getStackInSlot(1, 2).isOf(materialItem);
+    }
+
+    @Override
+    public ItemStack craft(CraftingRecipeInput input, RegistryWrapper.WrapperLookup lookup) {
+        // Get material
+        ItemStack materialItem = input.getStackInSlot(1, 0);
+        // Get Shulker
+        ItemStack shulkerStack = input.getStackInSlot(1, 1).copy();
+        // Apply material
+        ShulkerUtil.setMaterialForShulker(shulkerStack, materialItem);
         return shulkerStack;
     }
 
     @Override
     public boolean fits(int width, int height) {
-        return width >= 3 && height >= 3;
+        return width > 2 && height > 2;
     }
 
     @Override
     public ItemStack getResult(RegistryWrapper.WrapperLookup registriesLookup) {
-        return ItemStack.EMPTY;
+        return Items.SHULKER_BOX.getDefaultStack();
+    }
+
+    @Override
+    public boolean showNotification() {
+        return true;
     }
 
     @Override
@@ -91,21 +96,20 @@ public class ShulkerUpgradeRecipe implements CraftingRecipe {
         return CraftingRecipeCategory.MISC;
     }
 
-    /*
-    public static class Serializer implements RecipeSerializer<ShulkerUpgradeRecipe> {
-        public static final MapCodec<ShulkerUpgradeRecipe> CODEC = MapCodec.unit(ShulkerUpgradeRecipe::new);
-        public static final PacketCodec<RegistryByteBuf, ShulkerUpgradeRecipe> PACKET_CODEC = PacketCodec.unit(new ShulkerUpgradeRecipe());
+    @Override
+    public @Nullable Recipe<?> getPolymerReplacement(ServerPlayerEntity player) {
+        return PolymerRecipe.createCraftingRecipe(this);
+    }
 
-        @Override
-        public MapCodec<ShulkerUpgradeRecipe> codec() {
-            return CODEC;
+    public static class Serializer extends SpecialRecipeSerializer<ShulkerUpgradeRecipe> implements PolymerObject {
+
+        public Serializer(Factory<ShulkerUpgradeRecipe> factory) {
+            super(factory);
         }
 
         @Override
         public PacketCodec<RegistryByteBuf, ShulkerUpgradeRecipe> packetCodec() {
-            return PACKET_CODEC;
+            return null;
         }
     }
-
-     */
-} 
+}
