+++ src/main/resources/fabric.mod.json
@@ -13,7 +13,7 @@
 	},
 	"license": "CC0-1.0",
 	"icon": "assets/better-shulkers/icon.png",
-	"environment": "server",
+	"environment": "*",
 	"entrypoints": {
 		"main": [
 			"me.noramibu.bettershulkers.BetterShulkers"
@@ -23,12 +23,9 @@
 		"better-shulkers.mixins.json"
 	],
 	"depends": {
-		"fabricloader": ">=0.16.14",
+		"fabricloader": "*",
 		"minecraft": ">=1.21 <1.22",
 		"java": ">=21",
 		"fabric-api": "*"
-	},
-	"suggests": {
-		"another-mod": "*"
 	}
 }
